requirements.txt
requests
websocket-client
boto3
threaded
pysqlite3